feat(backup): 1. remove pervious implementation
