+++ src/rdsn/src/replica/backup/replica_backup_manager.h
@@ -23,6 +23,7 @@
 namespace dsn {
 namespace replication {
 
+// TODO(heyuchen): implement it
 class replica;
 class replica_backup_manager : replica_base
 {
@@ -30,20 +31,11 @@ public:
     explicit replica_backup_manager(replica *r);
     ~replica_backup_manager();
 
-    void on_clear_cold_backup(const backup_clear_request &request);
-    void start_collect_backup_info();
-
 private:
-    void clear_backup_checkpoint(const std::string &policy_name);
-    void send_clear_request_to_secondaries(const gpid &pid, const std::string &policy_name);
-    void background_clear_backup_checkpoint(const std::string &policy_name);
-    void collect_backup_info();
-
-    replica *_replica;
-    dsn::task_ptr _collect_info_timer;
-
     friend class replica;
     friend class replica_backup_manager_test;
+
+    replica *_replica;
 };
 
 } // namespace replication
