+++ src/rdsn/src/replica/backup/test/replica_backup_manager_test.cpp
@@ -21,27 +21,11 @@
 namespace dsn {
 namespace replication {
 
+// TODO(heyuchen): implement it
 class replica_backup_manager_test : public replica_test_base
 {
 public:
-    void clear_backup_checkpoint(const std::string &policy_name)
-    {
-        _replica->get_backup_manager()->clear_backup_checkpoint(policy_name);
-    }
 };
-
-TEST_F(replica_backup_manager_test, clear_cold_backup)
-{
-    std::string policy_name = "test_policy";
-
-    // create policy dir: <backup_dir>/backup.<policy_name>.*
-    std::string policy_dir = _replica->get_app()->backup_dir() + "/backup." + policy_name;
-    utils::filesystem::create_directory(policy_dir);
-
-    // clear policy dir
-    clear_backup_checkpoint(policy_name);
-    ASSERT_FALSE(utils::filesystem::directory_exists(policy_dir));
-}
 
 } // namespace replication
 } // namespace dsn
