+++ src/rdsn/src/replica/replica.h
@@ -71,10 +71,6 @@ class replica_split_manager;
 class replica_disk_migrator;
 class replica_follower;
 
-class cold_backup_context;
-typedef dsn::ref_ptr<cold_backup_context> cold_backup_context_ptr;
-struct cold_backup_metadata;
-
 namespace test {
 class test_checker;
 }
@@ -99,6 +95,16 @@ DSN_DECLARE_bool(reject_write_when_disk_insufficient);
 bool get_bool_envs(const std::map<std::string, std::string> &envs,
                    const std::string &name,
                    /*out*/ bool &value);
+
+// TODO(heyuchen): refactor it
+struct cold_backup_metadata
+{
+    int64_t checkpoint_decree;
+    int64_t checkpoint_timestamp;
+    std::vector<file_meta> files;
+    int64_t checkpoint_total_size;
+    DEFINE_JSON_SERIALIZATION(checkpoint_decree, checkpoint_timestamp, files, checkpoint_total_size)
+};
 
 struct deny_client
 {
@@ -163,7 +169,6 @@ public:
     void on_config_sync(const app_info &info,
                         const partition_configuration &config,
                         split_status::type meta_split_status);
-    void on_cold_backup(const backup_request &request, /*out*/ backup_response &response);
 
     //
     //    messages from peers (primary or secondary)
@@ -406,17 +411,6 @@ private:
                                            const std::string &chk_dir);
 
     /////////////////////////////////////////////////////////////////
-    // cold backup
-    virtual void generate_backup_checkpoint(cold_backup_context_ptr backup_context);
-    void trigger_async_checkpoint_for_backup(cold_backup_context_ptr backup_context);
-    void wait_async_checkpoint_for_backup(cold_backup_context_ptr backup_context);
-    void local_create_backup_checkpoint(cold_backup_context_ptr backup_context);
-    void send_backup_request_to_secondary(const backup_request &request);
-    // set all cold_backup_state cancel/pause
-    void set_backup_context_cancel();
-    void clear_cold_backup_state();
-
-    /////////////////////////////////////////////////////////////////
     // replica restore from backup
     bool read_cold_backup_metadata(const std::string &file, cold_backup_metadata &backup_metadata);
     // checkpoint on cold backup media maybe contain useless file,
@@ -561,8 +555,6 @@ private:
     primary_context _primary_states;
     secondary_context _secondary_states;
     potential_secondary_context _potential_secondary_states;
-    // policy_name --> cold_backup_context
-    std::map<std::string, cold_backup_context_ptr> _cold_backup_contexts;
     partition_split_context _split_states;
 
     // timer task that running in replication-thread
