+++ src/rdsn/src/replica/replica_init.cpp
@@ -360,8 +360,6 @@ error_code replica::init_app_and_prepare_list(bool create_new)
                     std::chrono::seconds(_options->checkpoint_interval_seconds),
                     get_gpid().thread_hash());
             }
-
-            _backup_mgr->start_collect_backup_info();
         }
     }
