+++ src/rdsn/src/replica/test/replica_test.cpp
@@ -32,13 +32,7 @@ namespace replication {
 class replica_test : public replica_test_base
 {
 public:
-    replica_test()
-        : pid(gpid(2, 1)),
-          _backup_id(dsn_now_ms()),
-          _provider_name("local_service"),
-          _policy_name("mock_policy")
-    {
-    }
+    replica_test() : pid(gpid(2, 1)) {}
 
     void SetUp() override
     {
@@ -46,11 +40,6 @@ public:
         stub->install_perf_counters();
         mock_app_info();
         _mock_replica = stub->generate_replica_ptr(_app_info, pid, partition_status::PS_PRIMARY, 1);
-
-        // set cold_backup_root manually.
-        // `cold_backup_root` is set by configuration "replication.cold_backup_root",
-        // which is usually the cluster_name of production clusters.
-        _mock_replica->_options->cold_backup_root = "test_cluster";
     }
 
     int get_write_size_exceed_threshold_count()
@@ -110,54 +99,6 @@ public:
         _app_info.is_stateful = true;
         _app_info.max_replica_count = 3;
         _app_info.partition_count = 8;
-    }
-
-    void test_on_cold_backup(const std::string user_specified_path = "")
-    {
-        backup_request req;
-        req.pid = pid;
-        policy_info backup_policy_info;
-        backup_policy_info.__set_backup_provider_type(_provider_name);
-        backup_policy_info.__set_policy_name(_policy_name);
-        req.policy = backup_policy_info;
-        req.app_name = _app_info.app_name;
-        req.backup_id = _backup_id;
-        if (!user_specified_path.empty()) {
-            req.__set_backup_path(user_specified_path);
-        }
-
-        // test cold backup could complete.
-        backup_response resp;
-        do {
-            _mock_replica->on_cold_backup(req, resp);
-        } while (resp.err == ERR_BUSY);
-        ASSERT_EQ(ERR_OK, resp.err);
-
-        // test checkpoint files have been uploaded successfully.
-        std::string backup_root = dsn::utils::filesystem::path_combine(
-            user_specified_path, _mock_replica->_options->cold_backup_root);
-        std::string current_chkpt_file =
-            cold_backup::get_current_chkpt_file(backup_root, req.app_name, req.pid, req.backup_id);
-        ASSERT_TRUE(dsn::utils::filesystem::file_exists(current_chkpt_file));
-        int64_t size = 0;
-        dsn::utils::filesystem::file_size(current_chkpt_file, size);
-        ASSERT_LT(0, size);
-    }
-
-    error_code test_find_valid_checkpoint(const std::string user_specified_path = "")
-    {
-        configuration_restore_request req;
-        req.app_id = _app_info.app_id;
-        req.app_name = _app_info.app_name;
-        req.backup_provider_name = _provider_name;
-        req.cluster_name = _mock_replica->_options->cold_backup_root;
-        req.time_stamp = _backup_id;
-        if (!user_specified_path.empty()) {
-            req.__set_restore_path(user_specified_path);
-        }
-
-        std::string remote_chkpt_dir;
-        return _mock_replica->find_valid_checkpoint(req, remote_chkpt_dir);
     }
 
     void force_update_checkpointing(bool running)
@@ -227,11 +168,6 @@ public:
     dsn::app_info _app_info;
     dsn::gpid pid;
     mock_replica_ptr _mock_replica;
-
-private:
-    const int64_t _backup_id;
-    const std::string _provider_name;
-    const std::string _policy_name;
 };
 
 TEST_F(replica_test, write_size_limited)
@@ -373,21 +309,6 @@ TEST_F(replica_test, update_allow_ingest_behind_test)
         ASSERT_EQ(get_allow_ingest_behind(), test.expected_value);
         reset_allow_ingest_behind();
     }
-}
-
-TEST_F(replica_test, test_replica_backup_and_restore)
-{
-    test_on_cold_backup();
-    auto err = test_find_valid_checkpoint();
-    ASSERT_EQ(ERR_OK, err);
-}
-
-TEST_F(replica_test, test_replica_backup_and_restore_with_specific_path)
-{
-    std::string user_specified_path = "test/backup";
-    test_on_cold_backup(user_specified_path);
-    auto err = test_find_valid_checkpoint(user_specified_path);
-    ASSERT_EQ(ERR_OK, err);
 }
 
 TEST_F(replica_test, test_trigger_manual_emergency_checkpoint)
