+++ src/rdsn/src/replica/test/replication_service_test_app.h
@@ -38,15 +38,4 @@ public:
     }
     virtual error_code start(const std::vector<std::string> &args) override;
     virtual dsn::error_code stop(bool /*cleanup*/) { return dsn::ERR_OK; }
-
-    // test for cold_backup_context
-    void check_backup_on_remote_test();
-    void read_current_chkpt_file_test();
-    void remote_chkpt_dir_exist_test();
-
-    void upload_checkpoint_to_remote_test();
-    void read_backup_metadata_test();
-    void on_upload_chkpt_dir_test();
-    void write_backup_metadata_test();
-    void write_current_chkpt_file_test();
 };
